+++ examples/demo/src/app.rs
@@ -5,7 +5,6 @@ use leptos_hotkeys::{
 };
 use leptos_meta::*;
 use leptos_router::*;
-use leptos_theme::{use_theme, Theme, ThemeProvider};
 
 #[component]
 pub fn Button(href: &'static str, children: Children) -> impl IntoView {
@@ -23,16 +22,14 @@ pub fn App() -> impl IntoView {
     provide_hotkeys_context(main_ref, false, scopes!("scope_a"));
 
     view! {
-        <Stylesheet id="leptos" href="/pkg/demo.css"/>
+        <Stylesheet id="leptos" href="/pkg/demo.css" />
         <main _ref=main_ref>
-            <ThemeProvider>
-                <Router>
-                    <Routes>
-                        <Route path="/" view=HomePage/>
-                        <Route path="/:else" view=ErrorPage/>
-                    </Routes>
-                </Router>
-            </ThemeProvider>
+            <Router>
+                <Routes>
+                    <Route path="/" view=HomePage />
+                    <Route path="/:else" view=ErrorPage />
+                </Routes>
+            </Router>
         </main>
     }
 }
@@ -43,7 +40,6 @@ fn HomePage() -> impl IntoView {
         "border border-1 border-[#1a1a1a] dark:border-[#fdfdfd] p-8 space-y-20 h-full";
     let current_scope = create_rw_signal("scope_a");
     let is_green = create_rw_signal(true);
-    let current_theme = use_theme();
 
     // leptos_hotkey specific logic
     fn go_to_link(key: &'static str, link: String, scope: &'static str) {
@@ -57,7 +53,7 @@ fn HomePage() -> impl IntoView {
     let HotkeysContext { toggle_scope, .. } = use_hotkeys_context();
 
     // global hotkeys
-    use_hotkeys!(("KEYs") => move |_| {
+    use_hotkeys!(("s") => move |_| {
         toggle_scope("scope_a".to_string());
         toggle_scope("scope_b".to_string());
 
@@ -69,7 +65,7 @@ fn HomePage() -> impl IntoView {
     });
 
     go_to_link(
-        "control+R",
+        "controlleft+r,controlright+r",
         "https://github.com/gaucho-labs/leptos_hotkeys".to_string(),
         "*",
     );
@@ -91,17 +87,8 @@ fn HomePage() -> impl IntoView {
         set_count.set(0);
     });
 
-    // scope_b related hotkeys
-    use_hotkeys!(("keyT", "scope_b") => move |_| {
-        if current_theme.get() == Theme::Light {
-            current_theme.set(Theme::Dark)
-        } else {
-            current_theme.set(Theme::Light)
-        }
-    });
-
     let a_ref = create_node_ref::<html::Div>();
-    use_hotkeys_ref!((a_ref, "digit6", "scope_a") => move |_| {
+    use_hotkeys_ref!((a_ref, "6", "scope_a") => move |_| {
         if is_green.get() {
             is_green.set(false)
         } else {
@@ -111,14 +98,14 @@ fn HomePage() -> impl IntoView {
 
     const BANANA: &str = "https://www.youtube.com/watch?v=N982sQcjsZI";
 
-    go_to_link("keyB+metaleft", BANANA.to_string(), "scope_b");
+    go_to_link("b+meta", BANANA.to_string(), "scope_b");
 
     view! {
         <div class="dark:bg-[#1a1a1a] bg-[#fdfdfd] dark:text-white flex justify-center h-screen py-20 w-full font-robotomono absolute">
 
             <div class="w-10/12 h-full flex flex-col space-y-20">
                 <div class="space-y-2 text-lg">
-                    <div class="flex space-x-8 flex items-end">
+                    <div class="flex space-x-8 items-end">
                         <Button href="https://github.com/gaucho-labs/leptos-hotkeys">
                             <p class="text-2xl">leptos_hotkeys</p>
                         </Button>
@@ -169,7 +156,7 @@ fn HomePage() -> impl IntoView {
                                 <p>scope_b</p>
                                 <div class="space-y-2">
                                     <p>press 'T' to switch themes</p>
-                                    <p>press "Cmd/Super/Win" + 'B'</p>
+                                    <p>press "Cmd/Super/Win"+ 'B'</p>
                                 </div>
 
                             </div>
@@ -201,7 +188,8 @@ fn ErrorPage() -> impl IntoView {
     let unknown = p_unknown();
 
     view! {
-        <div class="h-screen w-full flex flex-col items-center justify-center font-robotomono".to_string()>
+        <div class="h-screen w-full flex flex-col items-center justify-center font-robotomono"
+            .to_string()>
             <p class="">Unknown command: {unknown}</p>
         </div>
     }