+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## _April 18th, 2024_ - 0.2.0
+## _July 2nd, 2024_ - [0.2.2]
+- Fix bugs related to keycode management
+
+## _May 27th, 2024_ - [0.2.1]
+- Code movement, change to using keycodes
+
+
+- Fix right/left modifiers not being recognized.
+
+### Enhancements
+
+- Add `FromStr` implementation for `Hotkey` struct.
+
+## _April 18th, 2024_ - [0.2.0]
 
 ### Breaking changes
 
@@ -8,6 +21,7 @@
 - Remove `prelude` module.
 - Make ref nodes listen to `keydown` Keyboard events.
 - Remove uneeded features `csr` and `hydrate`.
+- Use `code` instead of `key` property for `KeyboardEvent`.
 
 ### Enhancements
 
@@ -15,9 +29,8 @@
 - Do not depend on `log` if `debug` feature is not enabled.
 - Relax dependency versions.
 - Add compatibility with Leptos stable.
-- Use `code` instead of `key` property for `KeyboardEvent`.
 
-## _February 27th, 2024_- 0.1.5
+## _February 27th, 2024_ - 0.1.5
 
 - Clean up macros
 
@@ -35,3 +48,6 @@
 
 - Elevate `leptos` to v0.6.5
 - Added `event.preventDefault()`.
+
+[0.2.1]: https://github.com/gaucho-labs/leptos-hotkeys/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/gaucho-labs/leptos-hotkeys/compare/b83afc96...v0.2.0